feat(Navbar): I have created a navbar to hold the Home page link buttons.
The Navbar now import the linkButton components I made. Can have upto 4 nav link currently. I need
to pass in nav objects that have a label and an onClick function. This will be passed so the button
can click it. The button are still independent of the navbar and can be used else where if need be.