feat(Middleware): Firebase interactions
I have initial loading of the recipeList working with the custom Middleware. I just dispatch the
buildList action from the RecipeList component. The middleware intercepts the actions and creates a
new action with the snap.val() as the recipeList property(could be payload to follow FSA). There are
dispatch calls to the async information in the store. Will allow me to have a spinner for loading.

closes #11