fix(AddRecipe): Capturing data
When I was adding a recipe before the new data was coming from the form itself and not the state of
the component. I also changed on EditRecipe the state property from recipe to data to be consistence
with AddRecipe and to mkae it easier to identify the new data.

closes #7
closes #13