refactor(AddRecipe): Between store and Firebase
Have converted the addRecipeToFirebase to just addRecipe action call. In the Firebase middleware I
check for that action. Then using the new stringifyRecipe funciton I convert the immutable map or JS
object to a plain JS object with the directions and ingredients as strings. That data is then sent
to Firebase via promise. However updating the redux store is not dependant on this. Just thinking is
there is a failure I will need to notify the user and remove the recipe they created.