refactor(Firebase Middleware): UpdateRecipe
Got updateRecipe switched over to the middleware. I created a new action for updateRecipe to just
update the entire thing by using delete on the oldRecipe and then addRecipe to add the new one. I
also create the development vs production configureStore.js files so that I slim down the store in
production. There is no more Firebase interactions in the actionCreators! The async actionCreators
may need renaming so that they are more gneral for all the DB interactions.

closes #6
closes #17