fix(Adding Recipe): Id Created date
Got the id and created date being adding to the recipe when being added. When AddRecipe is
constructed then I add a id and created date property to the state which gets used when passed
through to addRecipe action. I can change this to componentDidMount(). Need to fix the broken tests
and then can add redux devtools since the reducer will be pure! Well make sure.