feat(RecipeDetails):
Have got a basic layout done for the recipeDetails. I do not know how I am going to get the files
uploaded yet through this app. However I can upload to Cloudinary and store that url in the recipe
object, under  key. Will need to add in the edit recipe and add recipe a field for this
data. Will open an issue for that.