refactor(Redux): Multi Connected components -> a container
I have removed all the redux connections from all the child components on Main.js. Now Main is the
connected component to redux and the rest will be pass the props. Done by using this answer
(stackoverflow.com/questions/32370994/hot-to-pass-this-props-children). Uses React.Children.map to
map the this.props.children and a function as arguments to then use React.cloneElement(child,
{...this.props}.