feat(RecipeForm): Added Datepicker
Got the material-ui datepicker working in the inputForm. I had to chanfe the handleChange function
to accommadate the non-event. Meaning event is set to null and the second param is the data value.
This is not true for the textFields as tey fire synethic events, with a target.property and
target.value. I have live preview working for the Date picker and imageURL. Only problem was getting
the default date setup in EditRecipe. The date would never show so instead I just decided not to
worry about it, as the live preview works and if the user changes the date it will show in both
ssecitons.