feat(Login): Firebase auth
I am using Firebase to authenticate users so that when I login with me email and password I can
write to the database. However if the user uses Google then they do not get the admin status.
Installed Redux-form and will set it up for the addRecipe component as well.