refactor(Login): Changed to Middleware
This is where it belongs really. I used redux-form's handleSubmit which can take a function as a
parameter with values, dispatch for its arguments when called. I then created the required constants
and actions, using redux-action this time. I dispatch a login action passing in the values object as
the payload. This is then destructured to login the user. I have the same functionality at this
moment. Very poor. I need to change the other actionCreators to FSA with redux-actions. Change the
other forms to use redux-form. This removes the need of this.state. As well as that horrible
clearing function I hand to hack together. Now I can use redux-form resetForm to hand this very
easily!