refactor(redux-form): Start on addRecipe
I have got redux-form working now on the addRecipe component. There is a lot of hacky code to just
get this working. Therefore I need to go back when I am deleting the component state code, and
refactor some of this hackiness. Like the name of function x to handleSumbit. Dealing with formating
the data. Maybe it would be best to sneak in a reducer to fire off the format function. This would
have to respond to redux-form actions. I can get them from viewing the console logging actions.