refactor(Remove): Commented code out
I have started to remove the old way of handling the livePreview. I have commented out the state
creation in constructor, changed the functions around and therefore removed the x function. I need
to figure out a way to handle the formatting of the user inputs and build a newRecipe Map in the
handleSubmit function.