refactor(Redux-form): Refactored AddRecipe
After installing redux-form and using it to build the UserLogin form I figured out how to use it
with AddRecipe and remove all the local state in that component. Live preview now gets all the data
from spreading out a function call. That function is jsut hiding the mess that is checking is the
values are there to send into the format function. I was able to change the component to just a
function instead of the class. EditRecipe will be broken at this point since the InputForm is not
reading those props anymore. Instead I need to get the data like I do in AddRecipe by the
this.props.fields spread. I am going to open an issue to better name the handlePreview function so
that I can use it in the onSubmit function as well.