feat(Multiple):
I have made serveral changes on this commit. I added comments to the format.js file to help better
remember what I was thinking when making the functions. I added a setTimeout to the route change in
both add and edit Recipe components. This is to allow for cloudinary to upload the image. Not fool
proof and should fine better solution. I changed from the material-ui icons to font awesome. The
goal was to get the cog spinner when submitting the recipe, the delay part. However the submitting
props injectted from redux-form does not seem to be updating... There is now an if statement to
check for the img file from user before submitting the file to the server, and therefore cloudinary.
I believe that the upload will now work on heroku since I do not require the config file in
server.js. Linkbutton got a prop of type added. This allowed me to changed from onClick for the
submission to just using the form onSubmit. I just realized why I changed this in the first place
and will change back and swuash commit. I added a limit property to the bodyParser in server.js; this allows for up to 50mb (set by me) to be uploaded. Could be lowered for images only... The clearOne functionality of the InputForm component was
removed due to complications... lol