feat(Tons): Too big...
These have to stop involoving so much of a change to the codebase. I have changed both the onSubmit
functions from add/editRecipe so that the redux-form prop submitting will update. This allows me
to modify the font awesome icon I am using. On that note I have switched to react-fa from
react-fontawesome. This is due to the popularity and the fact it comes with the actually icons. This
required me to figure out how to self-host fonts and what webpack configuration was needed. In the
firebase middleware I have changed the next(action) call to happend after the promise returns from
hitting firebase. I currently have it in both success and failure functions. As it stands I belive I
have to conitune with adding the recipe locally and then jsut deleting it when on the recipeDetails
page. Which will require being changed from a function component to a extend React.component class.
This may fix the problem of setState during a render. I would like a better way, such as not even
adding the recipe and not changing the url path, but instead give the user a message about failure.
However the props were not updating in during the onSubmit function call. Meaning when I would
console.log out a flag I was changing inside of the middleware or even in a reducer the flag would
not change until after onSubmit() had completed. Which means the path changed.