feat(Submitting): Got failied submission working
I figured out how to stop the onSubmit function from continuing. I set the add or update action call
to the DBPromise which is the either the succes of fail function from the main firebase promise in
my middleware. I call .then() on this and check the returning action object. If there is an error
property set, which would happen on a fialed firebase promise. I then reject the onSubmit with
passing in a error. That error is spread out onto the component props as error and error.name Where
error is the _error property I declard on reject(). The errors.name is set to a message for the user
to receive. When the _error or erro prop is set to validation then it will be reflected in the text
input fields from material-ui. However when set to error is database then I wll have to display a
error component to the user. Which I have not created yet.