refactor(Image): Cloudinary to URI
Instead of send the base64 URI of the image to Cloudinary and saving the http url, I am now just
saving the URI as the image for the recipe. Then in recipeDetails the img tag can handle the base64
encoded string. I also have installed compression for express 4. I hope this will fix the gzip
compression that google page speed told me about.