style(Styling): Fuild design
I have started to work on a fuild mobile first design. I added in the navButton that I made in
CareerSmarts. I have the header as a fixed bar at the top of the screen. I would like to change to a
clear background when scroll. Just not sure how to do that yet. T use the flexbox rules I wrapped
the Logo and NavButton component in a div to handle the positioning in the FixedNavBar stylesheet.