feat(Login): More login options
I have added in the ability to login with Google, Github, Twitter, or Facebook. These are all done
with popups, which facebook warned about for smaller screens. I may have to change this to the
redirect version. I also got rid of fireproof as I learned that Firebase itself will return
promises. The only code that had to change was the getting the list. Instead on list.then() I had to
have list.once('value').then().  I started to spilt apart the format file as it was getting way to
big. The idea is to create a format folder and split up the functions into smaller files like
helpers.js and user.js for the user formatting functions. I added in Facebook's SDK thing which I do
not understand yet....