refactor(User): Interactions with Database
I have changed the stringify function to take in another param which is the firebaseRef.uid value. I
store it on the recipe as the authorID. Then using the rules for Firebase I say the only way to
write to the recipeList is to be logged in and no existing data, or have the current auth.uid == the
authorID of the recipe, or lastly be me the adminUser. I have refeactored some common code in the
middleware into a file in the same folder called helpers.js I have created a authenticate function
and a sendRecipe function. The goal is to change the entire middleware to be a switch statement with
simple function calls. This maybe be impractical in this situation. However I still with to keep up
the repeating code.